+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 is now tested against Gitlab CE and Gitlab EE.*
 
 FEATURES:
-* **New Resource:** `gitlab_project_push_rules` ([#163](https://github.com/terraform-providers/terraform-provider-gitlab/pull/163))
+* **New Resource:** `gitlab_project_push_rule` ([#163](https://github.com/terraform-providers/terraform-provider-gitlab/pull/163))
 * **New Resource:** `gitlab_deploy_key_enable` ([#176](https://github.com/terraform-providers/terraform-provider-gitlab/pull/176))
 * **New Resource:** `gitlab_project_share_group` ([#167](https://github.com/terraform-providers/terraform-provider-gitlab/pull/167))